Move TCP transport back to proper namespace
The TCP transport is technically still experimental, but conveying that by putting it in a different namespace is not optimal. Doing so makes it look out of place in the docs and results in a breaking change when it is eventually moved out of the experimental namespace.
